+++ app/features/instruction-program-mango/ui/GenericSpotMangoDetailsCard.tsx
@@ -2,23 +2,23 @@ import { Address } from '@components/common/Address';
 import { InstructionCard } from '@components/instruction/InstructionCard';
 
 import { BaseTable } from '@/app/shared/ui/Table';
-import { getPerpMarketFromInstruction } from '@explorer/decoder-mango';
+import { getSpotMarketFromInstruction } from '@explorer/decoder-mango';
 import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
 
-export function GenericPerpMngoDetailsCard(props: {
+export function GenericSpotMangoDetailsCard(props: {
     ix: TransactionInstruction;
     index: number;
     result: SignatureResult;
-    mangoAccountKeyLocation: number;
-    perpMarketKeyLocation: number;
+    accountKeyLocation: number;
+    spotMarketkeyLocation: number;
     title: string;
     innerCards?: JSX.Element[];
     childIndex?: number;
 }) {
-    const { ix, index, result, mangoAccountKeyLocation, perpMarketKeyLocation, title, innerCards, childIndex } = props;
-    const mangoAccount = ix.keys[mangoAccountKeyLocation];
-    const perpMarketAccountMeta = ix.keys[perpMarketKeyLocation];
-    const mangoPerpMarketConfig = getPerpMarketFromInstruction(ix, perpMarketAccountMeta);
+    const { ix, index, result, accountKeyLocation, spotMarketkeyLocation, title, innerCards, childIndex } = props;
+    const mangoAccount = ix.keys[accountKeyLocation];
+    const spotMarketAccountMeta = ix.keys[spotMarketkeyLocation];
+    const mangoSpotMarketConfig = getSpotMarketFromInstruction(ix, spotMarketAccountMeta);
 
     return (
         <InstructionCard
@@ -36,17 +36,17 @@ export function GenericPerpMngoDetailsCard(props: {
                 </BaseTable.Cell>
             </BaseTable.Row>
 
-            {mangoPerpMarketConfig !== undefined && (
+            {mangoSpotMarketConfig !== undefined && (
                 <BaseTable.Row>
-                    <BaseTable.Cell>Perp market</BaseTable.Cell>
-                    <BaseTable.Cell className="text-right">{mangoPerpMarketConfig.name}</BaseTable.Cell>
+                    <BaseTable.Cell>Spot market</BaseTable.Cell>
+                    <BaseTable.Cell className="text-right">{mangoSpotMarketConfig.name}</BaseTable.Cell>
                 </BaseTable.Row>
             )}
 
             <BaseTable.Row>
-                <BaseTable.Cell>Perp market address</BaseTable.Cell>
+                <BaseTable.Cell>Spot market address</BaseTable.Cell>
                 <BaseTable.Cell>
-                    <Address pubkey={perpMarketAccountMeta.pubkey} alignRight link />
+                    <Address pubkey={spotMarketAccountMeta.pubkey} alignRight link />
                 </BaseTable.Cell>
             </BaseTable.Row>
         </InstructionCard>
